requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-multipart>=0.0.6

# File management
gitpython>=3.1.44

# Web scraping
beautifulsoup4>=4.13.3
requests>=2.32.3

# Vector embeddings - compatible versions
transformers==4.37.1
sentence-transformers==2.2.2

# Database
sqlalchemy>=2.0.0
pymongo>=4.3.3

# Utilities
python-dotenv>=1.1.0

aiohttp==3.10.11
boto3==1.37.23
docx==0.2.4
networkx==3.4.2
numpy==2.2.4
playwright==1.51.0
pydantic_settings==2.8.1
pytest==8.3.5
weasyprint==65.0
httpx>=0.24.0
torch==2.4.1
markdown==3.5.1
APScheduler==3.10.1
